fix things a little
