+++ package.json
@@ -17,7 +17,7 @@
     "@types/react-dom": "^18.2.7",
     "@types/uuid": "^9.0.2",
     "d3": "^7.8.5",
-    "next": "^14.0.0",
+    "next": "^14.0.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.14.2",
@@ -29,7 +29,8 @@
   "scripts": {
     "dev": "next dev",
     "build": "next build",
-    "start": "next start"
+    "start": "next start",
+    "test": "next lint"
   },
   "eslintConfig": {
     "extends": [
