+++ tsconfig.json
@@ -21,9 +21,9 @@
     "jsx": "preserve",
     "baseUrl": "./",
     "paths": {
-      "@/*": [
-        "app/*"
-      ]
+      // "@/*": [
+      //   "/*"
+      // ]
     },
     "incremental": true,
     "plugins": [
