+++ types/domain.ts
